requirements.txt
openai-whisper
transformers
